The test for the iris functions of the IP Camera are in another class.
