+++ cameramodule/src/test/java/com/benine/ipcameracontrol/IpcameraTest.java
@@ -1,8 +1,6 @@
 package com.benine.ipcameracontrol;
 
 import static org.junit.Assert.assertEquals;
-import static org.junit.Assert.assertFalse;
-import static org.junit.Assert.assertTrue;
 
 import org.junit.Rule;
 import org.junit.Test;
@@ -79,119 +77,7 @@ public class IpcameraTest {
     assertEquals(0, res[0], 0.000001);
     assertEquals(180, res[1], 0.000001);
   }
-  
-  
-  
-  
-  @Test
-  public final void testSetAutoIrisOff() throws CameraConnectionException {
-    parameterList = new ArrayList<Parameter>();
-    parameterList.add(new Parameter("res", "1"));
-    parameterList.add(new Parameter("cmd", "#D30"));
-
-    final HttpRequest request = HttpRequest.request("/cgi-bin/aw_ptz")
-                                  .withQueryStringParameters(parameterList);
-    mockServerClient.when(request).respond(HttpResponse.response().withBody("d30"));
-
-    camera.getIris().setAutoIrisOn(false);
-    
-    mockServerClient.verify(request, VerificationTimes.once());
-  }
-  
-  @Test
-  public final void testSetAutoIrisOn() throws CameraConnectionException {
-    parameterList = new ArrayList<Parameter>();
-    parameterList.add(new Parameter("res", "1"));
-    parameterList.add(new Parameter("cmd", "#D31"));
-
-    final HttpRequest request = HttpRequest.request("/cgi-bin/aw_ptz")
-                                  .withQueryStringParameters(parameterList);
-    mockServerClient.when(request).respond(HttpResponse.response().withBody("d31"));
-
-    camera.getIris().setAutoIrisOn(true);
-    
-    mockServerClient.verify(request, VerificationTimes.once());
-  }
-  
-  @Test
-  public final void testIsAutoIrisOff() throws CameraConnectionException {
-    parameterList = new ArrayList<Parameter>();
-    parameterList.add(new Parameter("res", "1"));
-    parameterList.add(new Parameter("cmd", "#D3"));
-
-    final HttpRequest request = HttpRequest.request("/cgi-bin/aw_ptz")
-                                  .withQueryStringParameters(parameterList);
-    mockServerClient.when(request).respond(HttpResponse.response().withBody("d30"));
-
-    boolean res = camera.getIris().isAutoIrisOn();
-    
-    mockServerClient.verify(request, VerificationTimes.once());
-    assertFalse(res);
-  }
-  
-  @Test
-  public final void testIsAutoIrisOn() throws CameraConnectionException {
-    parameterList = new ArrayList<Parameter>();
-    parameterList.add(new Parameter("res", "1"));
-    parameterList.add(new Parameter("cmd", "#D3"));
-
-    final HttpRequest request = HttpRequest.request("/cgi-bin/aw_ptz")
-                                  .withQueryStringParameters(parameterList);
-    mockServerClient.when(request).respond(HttpResponse.response().withBody("d31"));
-
-    boolean res = camera.getIris().isAutoIrisOn();
-    
-    mockServerClient.verify(request, VerificationTimes.once());
-    assertTrue(res);
-  }
-  
-  @Test(expected = IpcameraConnectionException.class)
-  public final void testIsAutoIrisOnException() throws CameraConnectionException {
-    parameterList = new ArrayList<Parameter>();
-    parameterList.add(new Parameter("res", "1"));
-    parameterList.add(new Parameter("cmd", "#D3"));
-
-    final HttpRequest request = HttpRequest.request("/cgi-bin/aw_ptz")
-                                  .withQueryStringParameters(parameterList);
-    mockServerClient.when(request).respond(HttpResponse.response().withBody("p31"));
-
-    camera.getIris().isAutoIrisOn();
-  }
-  
-  @Test
-  public final void testSetIrisPosition() throws CameraConnectionException {
-    parameterList = new ArrayList<Parameter>();
-    parameterList.add(new Parameter("res", "1"));
-    parameterList.add(new Parameter("cmd", "#I80"));
-
-    final HttpRequest request = HttpRequest.request("/cgi-bin/aw_ptz")
-                                  .withQueryStringParameters(parameterList);
-    mockServerClient.when(request).respond(HttpResponse.response().withBody("iC80"));
-
-    camera.getIris().setIrisPos(80);
-    
-    mockServerClient.verify(request, VerificationTimes.once());
-  }
-  
-  @Test
-  public final void testGetIrisPosition() throws CameraConnectionException {
-    parameterList = new ArrayList<Parameter>();
-    parameterList.add(new Parameter("res", "1"));
-    parameterList.add(new Parameter("cmd", "#GI"));
-
-    final HttpRequest request = HttpRequest.request("/cgi-bin/aw_ptz")
-                                  .withQueryStringParameters(parameterList);
-    mockServerClient.when(request).respond(HttpResponse.response().withBody("giD421"));
-
-    int res = camera.getIris().getIrisPos();
-    
-    mockServerClient.verify(request, VerificationTimes.once());
-    
-    assertEquals(res, 3394, 0.000001);
-  }
-  
-  
-  
+   
   @Test
   public final void testGetStreamLink() {
     parameterList = new ArrayList<Parameter>();
@@ -201,7 +87,6 @@ public class IpcameraTest {
     String res = camera.getStreamLink();
     assertEquals(res, "http://127.0.0.1:9000/cgi-bin/mjpeg");
   }
-  
   
   @Test(expected = IpcameraConnectionException.class)
   public final void testNonExcistingIpAdres() throws CameraConnectionException {
